services/uporabniki/requirements.txt
Flask==2.2.3
Flask-PyMongo==2.3.0
bcrypt==4.0.1
pyjwt==2.6.0
